move keyboard listener into own file
